+++ packages/sdk/test/ui/use_watchable.test.tsx
@@ -162,6 +162,23 @@ describe('useWatchable', () => {
 
             mount(<Component />);
         });
+
+        it('will not accept undefined as the keys', () => {
+            // Suppress the console.error for cleaner test output.
+            jest.spyOn(console, 'error').mockImplementation(() => {});
+
+            const Component = () => {
+                // @ts-ignore
+                useWatchable(null, undefined);
+                return null;
+            };
+
+            expect(() => mount(<Component />)).toThrowError(
+                'Invalid call to useWatchable: keys cannot be undefined. ' +
+                    'Pass a key or array of keys corresponding to the model being watched as the ' +
+                    'second argument.',
+            );
+        });
     });
 
     describe('with several models', () => {